SDK-1539: Inline abstract builders
