+++ yoti-sdk-sandbox/src/main/java/com/yoti/api/client/sandbox/docs/request/SandboxTaskResults.java
@@ -34,7 +34,8 @@ public class SandboxTaskResults {
 
         private List<SandboxDocumentTextDataExtractionTask> documentTextDataExtractionTasks = new ArrayList<>();
 
-        private Builder() {}
+        private Builder() {
+        }
 
         public Builder withDocumentTextDataExtractionTask(SandboxDocumentTextDataExtractionTask textDataExtractionTask) {
             this.documentTextDataExtractionTasks.add(textDataExtractionTask);
