+++ yoti-sdk-sandbox/src/main/java/com/yoti/api/client/sandbox/docs/request/check/SandboxDocumentAuthenticityCheck.java
@@ -18,9 +18,10 @@ public class SandboxDocumentAuthenticityCheck extends SandboxDocumentCheck {
     /**
      * Builder for {@link SandboxDocumentAuthenticityCheck}
      */
-    public static class Builder extends SandboxDocumentCheckBuilder<Builder> {
+    public static class Builder extends SandboxDocumentCheck.Builder<Builder> {
 
-        private Builder() {}
+        private Builder() {
+        }
 
         @Override
         protected Builder self() {
