+++ yoti-sdk-sandbox/src/main/java/com/yoti/api/client/sandbox/docs/request/check/SandboxZoomLivenessCheckBuilder.java
@@ -5,9 +5,10 @@ import static com.yoti.api.client.spi.remote.util.Validation.notNull;
 import com.yoti.api.client.docs.DocScanConstants;
 import com.yoti.api.client.sandbox.docs.request.check.report.SandboxCheckReport;
 
-public class SandboxZoomLivenessCheckBuilder extends SandboxCheckBuilder<SandboxZoomLivenessCheckBuilder> {
+public class SandboxZoomLivenessCheckBuilder extends SandboxCheck.Builder<SandboxZoomLivenessCheckBuilder> {
 
-    SandboxZoomLivenessCheckBuilder() {}
+    SandboxZoomLivenessCheckBuilder() {
+    }
 
     @Override
     protected SandboxZoomLivenessCheckBuilder self() {
