+++ yoti-sdk-sandbox/src/main/java/com/yoti/api/client/sandbox/docs/request/check/report/SandboxBreakdown.java
@@ -49,7 +49,8 @@ public class SandboxBreakdown {
         private String result;
         private List<SandboxDetail> details = new ArrayList<>();
 
-        private Builder() {}
+        private Builder() {
+        }
 
         public Builder withSubCheck(String subCheck) {
             this.subCheck = subCheck;
